[svn] Added some debug printing and removed some. Improved documentation. Fixed URL.decode bug (it was a really silly thing). Changed how classes register themselves with their parents at init-time. Restored request.read[line] variable args. (Last checkin broke that).

+++ src/macros.h
@@ -53,7 +53,7 @@ THE SOFTWARE.
   #define log_debug(fmt, ...) fprintf(stderr, "DEBUG %s:%d: " fmt "\n", __FILE__, __LINE__, ##__VA_ARGS__)
   #define IFDEBUG(x) x
   #define assert_refcount(o, count_test) \
-    if(!((o)->ob_refcnt count_test)){ log_debug("assert_refcount(%ld %s)", (o)->ob_refcnt, #count_test); }\
+    if(!((o)->ob_refcnt count_test)){ log_debug("assert_refcount(%ld %s)", (long int)(o)->ob_refcnt, #count_test); }\
     assert((o)->ob_refcnt count_test)
 #else
   #define log_debug(fmt, ...) ((void)0)
